fix(dashboard-layout): fix imports, state model, and structure issues
- Import Fab from '../../base' instead of '@mui/material' (optional-peer guard)
- Import SxProps/Theme as 'import type' in WidgetPicker (erased at runtime)
- Replace useEffect state mirror with useRef-guarded two-dimension model:
  isSheetVisible is owned by DashboardLayout and resets to true only when
  isSidebarOpen transitions OFF→ON, allowing users to minimize the sheet
  while Edit Mode stays active (FAB appears as the re-open affordance)
- Move WidgetEmptyState body from index.tsx to WidgetEmptyState.tsx;
  index.tsx becomes a clean re-export barrel (consistent with other components)
- Remove duplicate WidgetPicker/DashboardLayout exports from index.tsx
  (already covered by index.ts)
