+++ src/custom/index.tsx
@@ -193,5 +193,3 @@ export * from './RJSFFormWrapper';
 export * from './ShareModal';
 export * from './UserSearchField';
 export * from './Workspaces';
-export * from './WidgetPicker';
-export * from './DashboardLayout';
